refactor: extract fat components into hooks (vote, global insight, filters)
Same "too much logic in one component" smell as the claim dialog, applied to
the three worst offenders the survey found. Logic moved verbatim into custom
hooks; components are now purely presentational. No behaviour change.

- serveur/vote-button → useVote: visitor status, cooldown timer, submit
  (captcha + 429 resync), success-head derivation.
- home/global-insight-section → useGlobalInsight: favorites-vs-manual selection,
  filters, aggregated `/global-stats` branch and per-server incremental cache.
- home/server-cards-section → useServerFilters: debounced search, category/
  language/edition filters, pagination, paginated SWR fetch.

Frontend builds clean; lint + typecheck pass.
